Update Bill Bundling
TODO: Gaslight player into thinking they counted correctly. If any bills in the bundled amount 30 are uncounted, then we show the unverified count to be equal to what the player thinks it should be meaning it adds up all the money that should be in each bundle so that it equals the expected total. Then the player is allowed to take whatever is leftover. So add a mechanic to keep money for yourself. If the cash is miscounted when bundling the actual total will be retained and during the deposit, the player will see the acutal total deposited with each bundle and with each bundle miscounted, the player takes extra time in a quick time event to add funds from his pocket to match the expected total.