Implement Combinational Mechanic for Select + Bundle.
TODO: update Select+Bundle so that when you press bundle after starting a Select+Bundle cash bundle, the remaining cash to fill the bundle is added and a new bundle is started if there is overflow. If you did not count the cash in the bundle, you will be gaslit into thinking that the cash is correctly counted. The count will appear accurate and will likely be miscounted, a new bundle as a result may be started for overflow even if there is no overflow. Conversely, if all bills are counted in the bundle, the correct amount will be counted and added and if there is overflow that will also be accurate.

If Bundle is pressed prior to a Select+Bundle combination, the Select+Bundle bill will be added to appear as accurately counted whether accounting for overflow or not. If the bundle is counted, it will also appear accurate and be accurate.

- TODO: add an indication for the user when Select is pressed to show that Bundle could also be pressed for a special action to count and start/add to a bundle.

- Update buttons to function based on "Press" rather than "Release".